move subcomponent to a subdirectory called profile
