+++ src/components/profile/UserPassword/index.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function UserPassword() {
+  return <div>UserPassword</div>;
+}
